pass and return records from sizes 1 to 100 by value There is some problem around passing records by value when I give the backend more type information. I'm having trouble tracking down the right code in the frontend.
It is important that the calls and functions agree on protocol,
even if that means records of certain sizes/types are passed
in registers, vs. others by pointer.

I'm very inclined to make the frontend make a simple consistent
decision to always pass by pointer, but let's see what it does
currently first.