I believe configure/autoconf does an adequate job of figuring out cc and cflags, so remove nearly all that mechanism here.
CC and CFLAGS will still be honored.

There is some supposition that you don't have a need
to set CC one way for m3cc and another way for e.g. m3core
and the m3 config files.
Since m3core/config files don't honor CC, that is reasonable.

We may or may not have a very pressing need to restore
this, so that we can remove -g on SPARC32_LINUX to save diskspace.
  I'd like to just replace that machine with one with more space.

The one part that is left is that for Solaris we
are willing to add /usr/sfw/bin or /opt/csw/gcc4/bin
to $PATH. We should probably put them at the end,
or only do so if cc is not available (gcc is buildable
with cc).