work in progress
I would like to make multiple passes over the m3cg intermediate form,
without "reparsing" it more than once. So, I would like to read it into
an array of structs. So, I would like the "marshalling" to be
more data driven/declarative, such as by using the existing
declarativeness but in a slightly more malleable form.
That is -- a "data" file that can be evaluated under the influence
of varying macros. Such as, one set of macros to declare struct
fields, and another set of macros to do the marshalling like currently.