for some reason, m3-convert.c needs to be combined with parse.c
in particular, on Solaris 2.10 with Sun CC, the value of
TREE_TYPE (expr) changes from the caller and convert.
The value in convert is garbage and it crashes.
This doesn't make sense, but I think it is ok to make the change
without understanding the rest of the problem.

If history preservation is paramount, we could #include it instead.

I didn't just merge the files, I also made edits to the convert function.
It had two temporaries that were identical to a paramter -- removed.
Current gcc 4.5 c-convert.c uses a switch instead of an if-latter, done.
removed support for vector/fixed/decimalfloat

in a few cases, e.g. INIT_FLOAT, LOAD_ADDRESS, use a local instead
of changing what is now a "persistant" member variable

fix the tracing during the handler calls
 (op vs. get_op())

support m3_break_lineno during the handler calls
  (far more interesting there, unless you want
  to debug the reading of the .mc file)

NOTE: the tracing during handler calls doesn't print the parameters;
that is only done in the reading pass. That can be fixed.

Generally the tracing isn't interesting in the reading pass,
only if there is a bug in reading.