don't inline anything, to fix: elego/m3msh/src/M3MiniShell.m3: In function 'M3MiniShell__ProcessParameters': elego/m3msh/src/M3MiniShell.m3:608:0: error: variable '_nonlocal_var_rec.188' might be clobbered by 'longjmp' or 'vfork'
very not good
I tried a fair number of other things and nothing both compiled that and compiled the rest
of the tree, at -O3.
-O2 was less of a problem.

It would *appear* that tree-nested copied volatility from the originals to the fields
of the nonlocal struct, but nevertheless, this error occurs.
(usually a warning, but I made it an error, we really don't want to let this occur)