Use 64K pages everywhere -- less target dependent code, basically.
The page size used to be target dependent when the garbage collector
interacted with virtual memory.
Now it is basically not machine dependent, except that on Windows
we use VirtualAlloc, which always allocates on 64K boundaries,
even if you ask for less than 64K, so allocation less than 64K
ends up completely wasting memory.

This could be viewed negatively as giving the Windows-dependent
value to all targets. But I think it is ok.