Set jumpbuf_align to 64 for all targets.
We'd really like to use 2 * BYTESIZE(INTEGER) for all targets
  except PPC_LINUX, but that isn't expressible. This is therefore
  as far as we know, incorrect for PA64_HPUX and SPARC64_SOLARIS,
  and non-ideal for PPC_LINUX, roughly the same as it ever was
  (well, the middle end got it right, but the runtime didn't,
  and they do need to match; so it was seemingly wrong, but
  somehow worked on PPC_LINUX)


In future, hopefully, compiler will inject the type instead
of depending on it being declared elsewhere in Modula-3.


In future, hopefully, this code will fall away on the majority
of platforms/users -- i.e. if we use the gcc stack walker.
But it will still stay around in some form for portability
to non-gcc platforms. So injecting the type is still desirable.


This removes lots of target-dependence.
This has the downside of introducing an extra 4 bytes of unnecessary
frame size on platforms where 4 byte alignment of jmpbuf is sufficient.
(Note that we have and continue to bloat the jmpbuf for other reasons.
  - sometimes it was sigjmpbuf, no longe
  - sometimes it is plain wrong, e.g. LINUXLIBC6, still
  - sometimes for hypothetical interop e.g. NT <=> Interix)


If RaiseActivation were one more or one less word,
or if the frame has an otherwise odd number of words,
the wastage should be removable by offseting the location
of RaiseActivation. That is, alignment should be achievable
by via hole, but via offseting the enclosing base and
possibly using the hole that that creates.


In either case, I'm inclined to go one step further and pass
an extern const to alloca and only put a pointer to the jmpbuf
in the "EF1", dramatically further eliminating target-dependence.
We are already calling pthread_getspecific and setjmp, so alloca
seems maybe a further reasonable deoptimization in an already slow path.
Perhaps we could somehow combine the calls (not easy, given
how setjmp and alloca are so special). This would presumably
also address the alignment problem -- assuming alloca knows
to return a pointer aligned enough for "anything", i.e. including
a jmpbuf.


There are still some platform-dependent Csetjmp.i3 files (NT)
in order to call longjmp instead of the more common _longjmp.
Testing, reading headers, and/or disasm will hopefully show
that such platforms have _longjmp and it is equivalent. Soon.