Documentation of ThreadPosix implementation as the light of understanding has finally shone on me and I wish to put down my observations for posterity.
Also an implementation of catching of SEGV and conversion into a
RuntimeError instead.  Mechanism is documented in ThreadPosix.m3:
it is a hybrid of the SIGCHLD handling and what ThreadPosix already
does to produce Thread.Alerted on a wake-up.  Mechanism has been
tested and appears to work: the former behavior was to quit execution
unconditonally; now it is instead a catcheable exception.