Initial checkin of Greg Nelson's Simplify theorem prover, used by the Modula-3 Extended Static Checker, the Java Extended Static Checker, and other programs.
I have verified that the code builds and runs under CM3 but I haven't
yet adjusted the build system to incorporate the code.

It's in the ESC subdirectory in the hope that HP will release the
rest of ESC soon.

Copyrights in the files may be misleading: all this code has long
since been released by DEC/Compaq/HP.

I downloaded it from sort.ucd.ie and made some very minor adjustments
to get it to build with the current CM3.