take a chance and let -O2 through
normally gcc does a two pass bootstrap,
build itself with some existing compiler, and then with itself
in the first pass, it omits -O2 because "some existing compiler"
often doesn't work very well

This change is risky. If it fails much, we can consider
undoing it, AND not using -disable-bootstrap.