split ScanCommandLine into ScanCommandLine1 and ScanCommandLine2 ScanCommandLine2 is most of the previous ScanCommandLine ScanCommandLine1 handles just -version and -help
Maybe call them ScanCommandLineForHelpAndUsage
and ScanCommadLine? or add a boolean parameter?
Or just call it earlier? This is a smaller/safer change.

The point is so that cm3 -version can be used to sniff
for -host, even when there is no (not yet) a cm3.cfg file.
This is useful e.g. in scripts/python/{pylib.py,boot2.sh}