some test cases for nested functions, except handlers, finally blocks, nested functions therein, and an initializer which it turns out is simple
There seems to be a matter where finally and/or except blocks are nested
but never take a static link, but other nested functions should take
an extra parameter for static link. I need to study M3x86.m3 more for how
it handle this, but it looks like maybe I should look examine
the function name to determine if it is an exception or finally handler,
and add static links to level>0 functions except for those. We'll see. Later.