implement FPU.scalb via FPU__ldexp, via a C wrapper, as is common now
I caught this because the declaration is wrong here,
int vs. INTEGER. The C backend causes an interesting
interop feature -- we actually get static checking
of our redeclarations, if #include is in the C code,
or if the compiler knows the function.

gcc does let us pass INTEGER to int w/o truncation warning,
so do that. Ok?