#ifdef __cplusplus typedef void* ADDRESS; #else typedef char* ADDRESS; #endif
and then change just a few occurences of ADDRESS to char*, i.e. for math
and then cast those to ADDRESS


in another great application my multi-pass infrastructure,
move declaration and initialization of segments (globals)
up to the start -- thereby mostly fixing the validity of the code as C++
(declaration would suffice, initialization could still be later)


There is still the matter that imported functions take
structs as "ADDRESS" but local functions take them as a pointer
to a struct with a specific size. I'll probably change that..
(Or did I already handle it? Anyway..progress toward valid C++,
not sure if I'm there yet, got side tracked fixing the rest of the system,
esp. the pp directory with lots of lex/flex output..)


Might want to go back to char* for ADDRESS.