http://www.gnu.org/software/gcc/gcc-4.8/changes.html
The struct reorg and matrix reorg optimizations
(command-line options -fipa-struct-reorg and
-fipa-matrix-reorg) have been removed. They
did not work correctly nor with link-time
optimization (LTO), hence were only applicable
to programs consisting of a single translation unit.


so remove them in our copies of older versions too.
Always looking for an excuse to remove gnarly code that
probably isn't used..that the owners removed it in newer
versions is a good reason..