use m3-sys/m3cc/I386_DARWIN to cross to AMD64_DARWIN, if it exists, and m3-sys/m3cc/I386_DARWIN-AMD64_DARWIN does not, and vice versa (actually, also should be willing to use *-I386_DARWIN to target AMD64_DARWIN, and vice versa, that is different, that is about chosing a cross compiler that has multiple targets, vs. this which is the native compiler, which has multiple targets)
reuses the pattern of InternalTargetCanRunOnHost and establishes
equivalency of many hosts/targets

Note that the config file elsewhere says to use "-m64" so that
cm3cg's defaulting to I386_DARWIN or AMD64_DARWIN is overriden.

This is somewhat esoteric stuff (cross builds), and might not be fully correct,
but probably is.