Favor Sun compilers over gcc and look for them in more places.
The motivation here is that I was getting a cm3cg that
wouldn't run, where it was just built, for lack of libgcc_s and libstdc++.
(No compiler is in $PATH. I ran gcc by full path, having found it
by looking around.)


The result here isn't actually ideal.
We now get a dependency instead of libCstd and libCrun, which
are present by default on the machine building cm3cg, but aren't
available across the board (checking just on opencsw).


What we really want to remove the C++ library dependency, or make
it static. There are switches to Sun cc/CC for this, but I'd
have to see if we can specify them in a correctly conditional way.


At least here, the thing we build on the host, runs on the host,
without messing with $PATH or $LD_LIBRARY_PATH.


Longer term, well, this all stinks ("this" being Unix build/runtime/setup
systems in genernal..)
I'd like to build a Solaris/x86 binary that works across a broad
selection of Solaris/x86 hosts. I'd like to build it on any of 5.8-11,
and have it work on any of 5.8-11. But build-host characteristics
are difficult to keep out.


The distant next best goal is to distribute portable source and have
it be recompiled on every host. Lame. We are getting there.
This is also a more reasonable goal to span Solaris, Darwin, Linux, NT, etc.


One might also imagine reverting m3cg to not use C++.
But you'd lose with gcc 4.8 anyway.
It requires C++.


Or maybe blame the g++ installation?
Maybe.