An almost solution to the problem of TextLiteral.T's having MaxBytes dependent on word size, making pickling of TextLiteral.T fail across different word sizes.
Hard code the fingerprint for the 64-bit case and translate it to
the one for 32-bit systems, in pickles. Also vice-versa.

The best long-term solution also involves making TextLiteral.MaxBytes
word-size-independent.  This could invalidate some preexisting pickles,
but only if they contain heap objects with a field or element with
declared static type TextLiteral.T, which seems quite unlikely.  If
such fields/methods are declared as TEXT, the actual allocated type of
their referents can be TextLiteral.T and all will work.