Win64: LAST(INTEGER) => LAST_INTEGER32 = 16_7FFFFFFF (right?) and use LAST_INTEGER32 instead. That is what was intended here.
Pending testing, replace this messy Modula-3 with the short clean C code already
neighboring. Really, the Modula-3 is doing some math gymnastics
that I can't decipher but surely is meant to be equivalent to the C code,
modulo possibly rounding.


The Modula-3 Win32 code is somewhat rife with the
questionable assumption that INTEGER is 32 bits.
Really, that isn't terrible, but we probably more code
that assumes INTEGER can hold pointers, so we will go with that.