get types from WinNT instead of local copy only trim one leading underscore for x86, not other architectures hardcode skipping of _real and not just _xmm (we pass _real in from config file) Handle _real and __real, _xmm and __xmm. The real symbols always have two underscores, and we trim one for x86. We could write the logic to check architecture and only skip two symbols instead of four, but this is ok.
I am somewhat keen on removing this entire utility, but it isn't trivial
at this point.

It is a glorified replacement for link /lib.

It produces .def files for us also.

It doesn't handle LTCG.

Ideally the .def files are produced by the backend
for any function in an capital I Interface.
And we can handle our C exports otherwise, if needed.

We should be filtering out data here too.
That has bitten us in the past, and would subsume _real/_xmm.

Error if we see multiple architectures.

And most importantly, allow IMAGE_FILE_MACHINE_AMD64.