Add Compiler.ThisEndian, as a CONST way of getting the endianness of the machine we run on.
NOTE: This starts out from Quake variable TARGET_ENDIAN, at the time the
compiler and m3core are compiled, when the interface Compiler is built.
This should be the host endianness when code that references Compiler.i3
is executed, which is what we want.  Confusing, but I think correct.

Bootstrapping this can be done with this possibly excessively pedantic
procedure:

1) Build and ship m3core.
2) Build, ship and install the compiler
3) Build libm3, which, in the devel_unicode branch, contains a use of
   the new facility, or will soon.

Step 1) might be unnecessary.