Initial changes to compiler and libraries to make WIDECHAR big enough for current Unicode.
Language:

LAST(WIDECHAR) = 16_10FFFF
Unicode escapes in char and text literals:
W'\U10FFFF' = VAL ( ORD ( 16_10FFFF ) , WIDECHAR )
Legal in nonwide literals, but value must be in range, e.g. \U0000FF.

Implementation:

BITSIZE(WIDECHAR) = 32

These are hard-coded in the compiler (m3front, m3cc(gcc-4.7)

Library changes (m3core) adapt on recompilation to the WIDECHAR range of
the compiler used to compile them.

Debugger adapts at its runtime to the WIDECHAR size of the code it is
debugging, but everything in the link closure needs to be recompiled
by a compiler with the same WIDECHAR size.

To bootstrap:
  1) Compile m3front and m3cc with an older compiler.
  2) Compile m3core with the compiler produced by step 1)